+++ engine/apps/schedules/tests/test_custom_on_call_shift.py
@@ -1297,7 +1297,7 @@ def test_get_oncall_users_for_empty_schedule(
     schedule = make_schedule(organization, schedule_class=OnCallScheduleCalendar)
     schedules = OnCallSchedule.objects.filter(pk=schedule.pk)
 
-    assert list(schedules.get_oncall_users()[schedule.pk]) == []
+    assert list(schedules.get_oncall_users()[schedule]) == []
 
 
 @pytest.mark.django_db
@@ -1413,7 +1413,7 @@ def test_get_oncall_users_for_multiple_schedules_emails_case_insensitive(
     oncall_users = schedules.get_oncall_users(events_datetime=events_datetime)
 
     assert len(oncall_users) == 1
-    assert list(oncall_users[schedule.pk]) == [user]
+    assert list(oncall_users[schedule]) == [user]
 
 
 @pytest.mark.django_db
@@ -1749,7 +1749,7 @@ def test_week_start_changed_daily_shift(
     on_call_shift.add_rolling_users(rolling_users)
 
     ical_data = on_call_shift.convert_to_ical()
-    expected_start = "DTSTART;VALUE=DATE-TIME:{}T000000Z".format(last_sunday.strftime("%Y%m%d"))
+    expected_start = "DTSTART:{}T000000Z".format(last_sunday.strftime("%Y%m%d"))
     assert expected_start in ical_data