+++ engine/apps/slack/scenarios/manage_responders.py
@@ -1,15 +1,13 @@
 import json
 import typing
 
-from apps.alerts.paging import DirectPagingAlertGroupResolvedError, check_user_availability, direct_paging, unpage_user
+from apps.alerts.paging import DirectPagingAlertGroupResolvedError, direct_paging, unpage_user, user_is_oncall
 from apps.slack.constants import DIVIDER
 from apps.slack.scenarios import scenario_step
 from apps.slack.scenarios.paging import (
-    DIRECT_PAGING_SCHEDULE_SELECT_ID,
     DIRECT_PAGING_USER_SELECT_ID,
+    _display_confirm_participant_invitation_view,
     _generate_input_id_prefix,
-    _get_availability_warnings_view,
-    _get_schedules_select,
     _get_select_field_value,
     _get_users_select,
 )
@@ -18,12 +16,10 @@ from apps.slack.types import Block, BlockActionType, EventPayload, ModalView, Pa
 
 if typing.TYPE_CHECKING:
     from apps.alerts.models import AlertGroup
-    from apps.schedules.models import OnCallSchedule
     from apps.slack.models import SlackTeamIdentity, SlackUserIdentity
     from apps.user_management.models import User
 
 MANAGE_RESPONDERS_USER_SELECT_ID = "responders_user_select"
-MANAGE_RESPONDERS_SCHEDULE_SELECT_ID = "responders_schedule_select"
 
 USER_DATA_KEY = "user"
 ALERT_GROUP_DATA_KEY = "alert_group_pk"
@@ -62,25 +58,23 @@ class ManageRespondersUserChange(scenario_step.ScenarioStep):
         selected_user = _get_selected_user_from_payload(payload)
         organization = alert_group.channel.organization
 
-        # check availability
-        availability_warnings = check_user_availability(selected_user)
-        if availability_warnings:
-            # display warnings and require additional confirmation
-            view = _get_availability_warnings_view(
-                availability_warnings,
-                organization,
-                selected_user,
-                ManageRespondersConfirmUserChange.routing_uid(),
-                json.dumps({USER_DATA_KEY: selected_user.id, ALERT_GROUP_DATA_KEY: alert_group.pk}),
+        # check if user is on-call
+        if not user_is_oncall(selected_user):
+            # display additional confirmation modal
+            private_metadata = json.dumps({USER_DATA_KEY: selected_user.id, ALERT_GROUP_DATA_KEY: alert_group.pk})
+            view = _display_confirm_participant_invitation_view(
+                ManageRespondersConfirmUserChange.routing_uid(), private_metadata
             )
+
             self._slack_client.views_push(trigger_id=payload["trigger_id"], view=view)
         else:
             try:
                 # no warnings, proceed with paging
                 direct_paging(
                     organization=organization,
-                    team=alert_group.channel.team,
                     from_user=slack_user_identity.get_user(organization),
+                    message=None,  # reuse the message from the original alert
+                    team=alert_group.channel.team,
                     users=[(selected_user, False)],
                     alert_group=alert_group,
                 )
@@ -111,8 +105,9 @@ class ManageRespondersConfirmUserChange(scenario_step.ScenarioStep):
         try:
             direct_paging(
                 organization=organization,
-                team=alert_group.channel.team,
                 from_user=slack_user_identity.get_user(organization),
+                message=None,  # reuse the message from the original alert
+                team=alert_group.channel.team,
                 users=[(selected_user, False)],
                 alert_group=alert_group,
             )
@@ -124,38 +119,6 @@ class ManageRespondersConfirmUserChange(scenario_step.ScenarioStep):
             trigger_id=payload["trigger_id"],
             view=view,
             view_id=payload["view"]["previous_view_id"],
-        )
-
-
-class ManageRespondersScheduleChange(scenario_step.ScenarioStep):
-    """Handle schedule selection in responders modal."""
-
-    def process_scenario(
-        self,
-        slack_user_identity: "SlackUserIdentity",
-        slack_team_identity: "SlackTeamIdentity",
-        payload: EventPayload,
-    ) -> None:
-        alert_group = _get_alert_group_from_payload(payload)
-        selected_schedule = _get_selected_schedule_from_payload(payload)
-        organization = alert_group.channel.organization
-
-        try:
-            direct_paging(
-                organization=organization,
-                team=alert_group.channel.team,
-                from_user=slack_user_identity.get_user(organization),
-                schedules=[(selected_schedule, False)],
-                alert_group=alert_group,
-            )
-            view = render_dialog(alert_group)
-        except DirectPagingAlertGroupResolvedError:
-            view = render_dialog(alert_group, alert_group_resolved_warning=True)
-
-        self._slack_client.views_update(
-            trigger_id=payload["trigger_id"],
-            view=view,
-            view_id=payload["view"]["id"],
         )
 
 
@@ -195,12 +158,12 @@ def render_dialog(alert_group: "AlertGroup", alert_group_resolved_warning=False)
                 Block.Section,
                 {
                     "type": "section",
-                    "text": {"type": "mrkdwn", "text": f":bust_in_silhouette: *{user.name or user.username}*"},
+                    "text": {"type": "mrkdwn", "text": f":bust_in_silhouette: *{user['name'] or user['username']}*"},
                     "accessory": {
                         "type": "button",
                         "text": {"type": "plain_text", "text": "Remove", "emoji": True},
                         "action_id": ManageRespondersRemoveUser.routing_uid(),
-                        "value": str(user.pk),
+                        "value": str(user["id"]),
                     },
                 },
             ),
@@ -220,16 +183,11 @@ def render_dialog(alert_group: "AlertGroup", alert_group_resolved_warning=False)
             ),
         ]
 
-    # Show user and schedule dropdowns
+    # Show user dropdown
     input_id_prefix = _generate_input_id_prefix()
-    blocks += [
+    blocks.append(
         _get_users_select(alert_group.channel.organization, input_id_prefix, ManageRespondersUserChange.routing_uid())
-    ]
-    blocks += [
-        _get_schedules_select(
-            alert_group.channel.organization, input_id_prefix, ManageRespondersScheduleChange.routing_uid()
-        )
-    ]
+    )
 
     view: ModalView = {
         "type": "modal",
@@ -262,17 +220,6 @@ def _get_selected_user_from_payload(payload: EventPayload) -> "User":
     return User.objects.get(pk=selected_user_id)
 
 
-def _get_selected_schedule_from_payload(payload: EventPayload) -> "OnCallSchedule":
-    from apps.schedules.models import OnCallSchedule
-
-    input_id_prefix = json.loads(payload["view"]["private_metadata"])["input_id_prefix"]
-    selected_schedule_id = _get_select_field_value(
-        payload, input_id_prefix, ManageRespondersScheduleChange.routing_uid(), DIRECT_PAGING_SCHEDULE_SELECT_ID
-    )
-
-    return OnCallSchedule.objects.get(pk=selected_schedule_id)
-
-
 def _get_alert_group_from_payload(payload: EventPayload) -> "AlertGroup":
     from apps.alerts.models import AlertGroup
 
@@ -291,12 +238,6 @@ STEPS_ROUTING: ScenarioRoute.RoutingSteps = [
         "payload_type": PayloadType.VIEW_SUBMISSION,
         "view_callback_id": ManageRespondersConfirmUserChange.routing_uid(),
         "step": ManageRespondersConfirmUserChange,
-    },
-    {
-        "payload_type": PayloadType.BLOCK_ACTIONS,
-        "block_action_type": BlockActionType.STATIC_SELECT,
-        "block_action_id": ManageRespondersScheduleChange.routing_uid(),
-        "step": ManageRespondersScheduleChange,
     },
     {
         "payload_type": PayloadType.BLOCK_ACTIONS,